Refactor path operations to use crud functions.
This closes #30

- Move all the database related functions and interactions
from the path operations to a dedicated crud file and make
function calls to interact with the database.

Refactor the get_users path operation.

- Move the factionality that interacts with the database to
a new function in the crud file.

Refactor the register_user path operation.

- Move the functionality that interacts with the database
to the crud.py file.

Refactor path operations to use crud functions.

- Move all the database related functions and interactions
from the path operations to a dedicated crud file and make
function calls to interact with the database.

Refactor the get_users path operation.

- Move the factionality that interacts with the database to
a new function in the crud file.

Refactor the register_user path operation.

- Move the functionality that interacts with the database
to the crud.py file.